+++ src/main/java/io/github/itzispyder/clickcrystals/mixins/MixinCamera.java
@@ -47,7 +47,7 @@ public abstract class MixinCamera implements Global {
     @ModifyArgs(method = "update", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/render/Camera;setRotation(FF)V"))
     private void onUpdateSetRotationArgs(Args args) {
         FreeLook freeLook = Module.get(FreeLook.class);
-        if (freeLook.isEnabled() && mc.options.getPerspective() == freeLook.perspectivePoint.getVal().getPerspective()) {
+        if (freeLook.isEnabled() && mc.options.getPerspective() == freeLook.perspectivePoint.getVal()) {
             args.set(0, freeLook.cY);
             args.set(1, freeLook.cP);
         }
